% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdf.R
\name{pdfinfo}
\alias{pdffonts}
\alias{pdfimages}
\alias{pdfinfo}
\alias{pdfseparate}
\alias{pdftohtml}
\alias{pdftops}
\alias{pdftotext}
\alias{pdfunite}
\alias{poppler}
\alias{xpdf}
\title{XPDF utilities}
\usage{
pdfinfo(file, args = "")

pdftotext(file, args = "")

pdftohtml(file, args = "")

pdftops(file, args = "")

pdfimages(file, args = "")

pdffonts(file, args = "")

pdfseparate(file, args = "")

pdfunite(file, args = "")
}
\arguments{
\item{file}{path or name of the PDF file}

\item{args}{additional arguments passed to the cmd util}
}
\description{
Dump wrappers around xpdf command line utilites.
}
\examples{
setwd(tempdir())
download.file("https://cran.r-project.org/doc/manuals/r-release/R-intro.pdf", "R-intro.pdf")
pdfinfo("R-intro.pdf")
pdffonts("R-intro.pdf")

# convert to text
pdftotext("R-intro.pdf")
browseURL("R-intro.txt")

# convert to html
pdftohtml("R-intro.pdf")
browseURL("R-intro.html")
}

