% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdf.R
\name{pdftotext}
\alias{pdfinfo}
\alias{pdftotext}
\alias{xpdf_version}
\title{Extract text from PDF file}
\usage{
pdftotext(input, output = "out.txt", firstpage = 1, lastpage = 0,
  layout = FALSE, table = FALSE, lineprinter = FALSE, raw = FALSE,
  fixed = 0, linespacing = 0, clip = FALSE, enc = "", eol = "",
  nopgbrk = FALSE, opw = "", upw = "", quiet = FALSE, cfg = "")

pdfinfo(input, enc = "", opw = "", upw = "", cfg = "")

xpdf_version()
}
\arguments{
\item{input}{pdf path or filename}

\item{output}{txt path or filename}

\item{firstpage}{first page to convert}

\item{lastpage}{last page to convert}

\item{layout}{maintain original physical layout}

\item{table}{similar to \code{layout} but optimized for tables}

\item{lineprinter}{use strict fixed-pitch/height layout}

\item{raw}{keep strings in content stream order}

\item{fixed}{assume fixed-pitch (or tabular) text}

\item{linespacing}{fixed line spacing for LinePrinter mode}

\item{clip}{separate clipped text}

\item{enc}{output text encoding name}

\item{eol}{output end-of-line convention (unix, dos, or mac)}

\item{nopgbrk}{don't insert page breaks between pages}

\item{opw}{owner password (for encrypted files)}

\item{upw}{user password (for encrypted files)}

\item{quiet}{don't print any messages or errors}

\item{cfg}{configuration file to use in place of .xpdfrc}
}
\description{
An R interface to the \code{pdftotext} utility from \code{xpdf}.
}
\examples{
download.file("https://cran.r-project.org/doc/manuals/r-release/R-exts.pdf", "R-exts.pdf")
pdfinfo("R-exts.pdf")
pdftotext("R-exts.pdf")
}

